aptean-hackathon/backend/requirements.txt
flask
flask-cors
sentence-transformers==2.2.2
transformers==4.35.0
torch>=1.12.0
scikit-learn
python-dotenv
uvicorn
faiss-cpu; platform_system != "Windows"
